File re-org, better comments
