+++ logger.go
@@ -14,11 +14,16 @@ var (
 	loggingWriter  io.Writer
 )
 
+// SetLogging adjusts the configuration for logging. You can enable
+// and disable the logging here. By default, logging is disabled.
+//
+// Most calls to this function will be called like SetLogging(true, os.Stdout)
 func SetLogging(enabled bool, writer io.Writer) {
 	loggingEnabled = enabled
 	loggingWriter = writer
 }
 
+// printQuery prints a query if the user has enabled logging
 func printQuery(query string, args ...interface{}) {
 	if loggingEnabled {
 		for i, arg := range args {
@@ -28,8 +33,40 @@ func printQuery(query string, args ...interface{}) {
 	}
 }
 
+// pointerToLogString converts a value pointer to the string
+// that should be logged for it
 func pointerToLogString(pointer interface{}) string {
 	switch v := pointer.(type) {
+	case *string:
+		return "'" + *v + "'"
+	case *float32:
+		return strconv.FormatFloat(float64(*v), 'f', -1, 32)
+	case *float64:
+		return strconv.FormatFloat(*v, 'f', -1, 64)
+	case *bool:
+		return strconv.FormatBool(*v)
+	case *int:
+		return strconv.Itoa(*v)
+	case *int8:
+		return strconv.FormatInt(int64(*v), 10)
+	case *int16:
+		return strconv.FormatInt(int64(*v), 10)
+	case *int32:
+		return strconv.FormatInt(int64(*v), 10)
+	case *int64:
+		return strconv.FormatInt(*v, 10)
+	case *uint:
+		return strconv.FormatUint(uint64(*v), 10)
+	case *uint8:
+		return strconv.FormatUint(uint64(*v), 10)
+	case *uint16:
+		return strconv.FormatUint(uint64(*v), 10)
+	case *uint32:
+		return strconv.FormatUint(uint64(*v), 10)
+	case *uint64:
+		return strconv.FormatUint(*v, 10)
+	case *time.Time:
+		return "'" + (*v).Format(time.RFC3339) + "'"
 	case *null.Int:
 		if v.Valid {
 			return strconv.FormatInt(v.Int64, 10)
@@ -55,27 +92,6 @@ func pointerToLogString(pointer interface{}) string {
 			return "'" + v.Time.Format(time.RFC3339) + "'"
 		}
 		break
-	case *string:
-		return "'" + *v + "'"
-	case *float32:
-	case *float64:
-		return strconv.FormatFloat(float64(*v), 'f', -1, 64)
-	case *bool:
-		return strconv.FormatBool(*v)
-	case *int:
-	case *int8:
-	case *int16:
-	case *int32:
-	case *int64:
-		return strconv.FormatInt(int64(*v), 10)
-	case *uint:
-	case *uint8:
-	case *uint16:
-	case *uint32:
-	case *uint64:
-		return strconv.FormatUint(uint64(*v), 10)
-	case *time.Time:
-		return "'" + (*v).Format(time.RFC3339) + "'"
 	default:
 		return fmt.Sprintf("%v", v)
 	}
